Added basic install capabilities
In "config.ini" you can specificy an "install_do" directory where the generated "def" folder should be moved to once built.
This saves the step of cut and pasting it manually.

This commit adds an example ini file, which you should copy and rename to "config.ini" and then adjust the paths accordingly.